Refactor the Design
