+++ src/atomdb/adapterdb/AdapterDB.h
@@ -13,6 +13,8 @@
 #include <vector>
 
 #include "AtomDB.h"
+#include "BoundedSharedQueue.h"
+#include "DatabaseMappingStrategy.h"
 #include "DatabaseTypes.h"
 #include "JsonConfig.h"
 #include "Utils.h"
@@ -22,18 +24,17 @@ using namespace db_adapter;
 
 namespace atomdb {
 
-enum class AdapterDbType { Postgres, Mork };
+enum class AdapterDbType { Postgres };
 
-inline AdapterDbType parse_adapter_db_type(const std::string& value) {
+inline AdapterDbType parse_adapter_db_type(const string& value) {
     if (value == "postgres") return AdapterDbType::Postgres;
-    if (value == "mork") return AdapterDbType::Mork;
     RAISE_ERROR("Unsupported adapterdb.type: " + value);
 }
 
 class AdapterDB : public AtomDB {
    public:
     explicit AdapterDB(const JsonConfig& config);
-    AdapterDB(const JsonConfig& config, std::shared_ptr<AtomDB> backend);  // for testing
+    AdapterDB(const JsonConfig& config, shared_ptr<AtomDB> backend);  // for testing
     ~AdapterDB() override;
 
     static string MONGODB_ADAPTER_COLLECTION_NAME;
@@ -104,6 +105,7 @@ class AdapterDB : public AtomDB {
     size_t atom_count() const override;
 
    private:
+    AdapterDbType adapter_type;
     JsonConfig config;
     shared_ptr<AtomDB> atomdb_backend;
     atomic<bool> backend_ready{false};
@@ -142,6 +144,9 @@ class AdapterDB : public AtomDB {
      * vector of strings.
      */
     vector<string> get_mapping_file_contents() const;
+
+    shared_ptr<DatabaseMappingStrategy> create_mapping_strategy(
+        shared_ptr<BoundedSharedQueue> queue) const;
 
     /**
      * @brief Runs the full adapter pipeline (orchestrator -> persister) synchronously.
