--- src/db_adapter/postgres/BUILD
@@ -1,30 +1,0 @@
-load("@rules_cc//cc:cc_library.bzl", "cc_library")
-
-package(default_visibility = ["//visibility:public"])
-
-cc_library(
-    name = "postgres_lib",
-    includes = ["."],
-    deps = [
-        ":postgres_wrapper",
-    ],
-)
-
-cc_library(
-    name = "postgres_wrapper",
-    srcs = ["PostgresWrapper.cc"],
-    hdrs = ["PostgresWrapper.h"],
-    includes = ["."],
-    linkopts = [
-        "-L/usr/local/lib",
-        "-lpqxx",
-        "-lpq",
-    ],
-    deps = [
-        "//commons:commons_lib",
-        "//db_adapter:bounded_shared_queue",
-        "//db_adapter:database_connection",
-        "//db_adapter:database_mapper",
-        "//db_adapter:database_wrapper",
-    ],
-)
