+++ src/db_adapter/sql/SqlWrapper.cc
@@ -1,0 +1,6 @@
+#include "SqlWrapper.h"
+
+using namespace db_adapter;
+
+SQLWrapper::SQLWrapper(DatabaseConnection& db_client, shared_ptr<DatabaseMapper> mapper)
+    : DatabaseWrapper(db_client, mapper) {}
